Tranzila: Attempt at supporting stupid API.
I don't take pride in this code. In fact I'd rather have it committed to
my memory as a bad dream. Sadly that is not the case.

In order to properly show RTL and LTR custom templates, which are
required if we want our sites to look more recent than 1996, we are
forced to use `template=custom` for LTR and `template=custom_he` for
Hebrew RTL template. It does, however, remain to be seen how complex
will this get once we have to do Arabic or some other RTL language.

Fixes with this patch are lacking logic and are hacky at best. When
instructing Tranzila to show RTL template we have to specify `lang=heb`
parameter in URL while sending the rest of the data through POST. Due to
lacking documentation we don't know if POST parameter with same name is
ignored or overridden (or overrides GET). When showing LTR template,
POST parameter is used and we omit specifying it 2 times.

It's important to note that parameter value is indeed `heb`. It's not a
typo. Don't try to fix it or you'll break the code. Another important
note is that when sending language through POST parameters language code
for Hebrew is `il`. Why this is like that I don't know. I can only
assume developers at Tranzila are smoking bath salts and are high most
of the time.

If this code is broken please don't ask me to fix it. Because I won't.
Ever. Just find a different payment gateway and move there regardless of
how much money would that cost you. Trust me, it's for the sake of your
and sanity of others around you.