Shop: Make score for search results be a function of matched words.
Up to this point search would award fixed number of points to the total
matching score of the item. This has a flaw where if more words are
specified the more items system will match and then sort by the highest
score.

New approach is to reserve points for each part of the item being searched
and only assign portion of score based on how many words are found in
the search portion. If item contains 100% of searched words, we increase
score by 100% of the reserved points.