Tranzila: Another fix for compabitility with stupid service.
Due to absolutely moronic choices in development and constant naming we
have to change `tranmode` parameter yet again. Apparently if this
parameter has `K` in it, it's meant to be used with tokens but it works
either way, except when it doesn't. Sadly this change probably breaks
some sites and there's very little I can do as Tranzila has instructed
us to change from `AK` to `A` and it will magically work, until next
time it doesn't.