Tranzila: Fix their view of mobile support.
It's really sad that in this day an age we should resolve to hacks of
this kind. In general we don't have an option as we are not allowed to
change anything inside of `iframe` and Tranzila developers never heard
of testing. Solution therefore is to increase `iframe` size and scale it
down. Phones today have high enough pixel density for this to not become
too big of an issue.