Refactor property adding code.
It's obvious that current code for adding property as title fragment
will be repeated a lot. To avoid this `add_property_to_title` method was
added to `ItemManager` class. This will allow individual handlers to
easily specify which properties can be added as title fragment and in
this way signifficantly simplify code.