Window system: Fixed issues with multiple connected load events.
This issue was subtly causing image upload to only work the first time.
Issue originated upon opening image upload window which caused another
event listened to be connected to the same `iframe`. This resulted in
multiple false triggering of the same and in doing so caused triggering
of handlers which failed silently.